Move volvooncall base entity to separate module
