+++ homeassistant/components/volvooncall/switch.py
@@ -12,8 +12,9 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import VolvoEntity, VolvoUpdateCoordinator
+from . import VolvoUpdateCoordinator
 from .const import DOMAIN, VOLVO_DISCOVERY_NEW
+from .entity import VolvoEntity
 
 
 async def async_setup_entry(
